requirements.txt
google-genai>=0.4.0
pandas>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
datasets>=2.14.0